emo_outlet_api/requirements.txt
# 情绪出口 API 依赖清单

# === Web 框架 ===
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# === 数据库 ===
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0  # 开发环境 SQLite
aiomysql>=0.2.0    # 生产环境 MySQL
alembic>=1.13.0

# === 认证安全 ===
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# === AI / LLM ===
openai>=1.12.0

# === 配置 ===
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# === 工具 ===
python-multipart>=0.0.6
httpx>=0.26.0
